+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v12.4.15
+# New Features
+- *Find Actor in Scenes*: You can now get a report of all scenes in your world that contain a reference to the selected actor.
+  - Right-click an actor to find all scenes for that single actor.
+  - Right-click an actor folder to generate a report for all contained actors. (recursive)
 # v12.4.14
 # New Features
 - System Identifier Normalization fuction
@@ -659,6 +664,7 @@
 - Fixed missing images.
 # v1.0.0: First test release
 - Included all content converted from v9.
+